Some refactoring, less hacky but still flickers
