+++ src/components/mobile/mobileTimeslotContent.tsx
@@ -4,8 +4,8 @@ import { useState, useContext } from "react";
 import Horse from "../../images/horseRider.svg";
 import Dude from "../../images/person.svg";
 import Bookmark from "../../images/bookmark.svg";
-import OnSlide from "../../images/onslider.png";
-import OffSlide from "../../images/offslider.png";
+import OnSlide from "../../images/OnSlider.png";
+import OffSlide from "../../images/OffSlider.png";
 import UserContext from "../../userContext";
 import MobileTimeSlotConfirmation from "./mobileTimeslotConfirmation";
 import TimeslotSuccess from "../popup/timeslotSuccess";
