+++ src/components/popup/timeslotPopup.tsx
@@ -1,11 +1,12 @@
-import React, { useState, useEffect, useContext } from "react";
+import React, { useState, useLayoutEffect, useContext, useMemo } from "react";
 import styled from "styled-components";
+import { DataStore } from "@aws-amplify/datastore";
 import x from "../../images/X.svg";
 import { PopupDiv, PopupBox, X, CancelBtn, SaveBtn } from "../styledComponents";
 import Monthly from "../monthlyView";
 import AptInfo from "../appointmentInfo";
 import Timeslots from "./timeslots";
-import { LazyTimeslot } from "../../models";
+import { User, LazyUser, LazyBooking, LazyTimeslot } from "../../models";
 import TimeslotConfirmation from "./timeslotConfirmation";
 import TimeslotSuccess from "./timeslotSuccess";
 import UserContext from "../../userContext";
@@ -83,7 +84,8 @@ export default function Popup({
   const [realUser] = currentUser;
   const { userType, id } = realUser;
   const [bookable, setBookable] = useState<TsData[]>([]);
-  const [selected, setSelected] = useState<LazyTimeslot>();
+  const [volunteerBookings, setVolBookings] = useState<LazyUser[]>([]);
+  const [riderBookings, setRidBookings] = useState<LazyUser[]>([]);
 
   const options: Intl.DateTimeFormatOptions = {
     weekday: "long",
@@ -91,22 +93,21 @@ export default function Popup({
     day: "numeric",
   };
   const formattedDate = date.toLocaleDateString("en-US", options);
+  const getSelected = () =>
+    timeslots.find((timeslot) => {
+      if (timeslot.startTime) {
+        const time = timeslot.startTime.split(":");
+        return (
+          Number(time[0]) === date.getHours() &&
+          Number(time[1]) === date.getMinutes()
+        );
+      }
+      return false;
+    });
+  const selected = useMemo(() => getSelected(), [popup]);
 
-  useEffect(() => {
+  useLayoutEffect(() => {
     const ts: TsData[] = [];
-    const getSelected = () => {
-      const sel = timeslots.find((timeslot) => {
-        if (timeslot.startTime) {
-          const time = timeslot.startTime.split(":");
-          return (
-            Number(time[0]) === date.getHours() &&
-            Number(time[1]) === date.getMinutes()
-          );
-        }
-        return false;
-      });
-      setSelected(sel);
-    };
 
     const fetchBookable = async () => {
       if (timeslots.length > 0) {
@@ -165,8 +166,43 @@ export default function Popup({
       }
       setBookable(ts);
     };
-    getSelected();
+    const getUsers = async (bookings: LazyBooking[]) => {
+      const volUsers: User[] = [];
+      const ridUsers: User[] = []; // eslint-disable-next-line no-restricted-syntax
+      for await (const booking of bookings) {
+        if (booking.date) {
+          if (selected) {
+            if (
+              Number(booking.date.substring(0, 4)) === date.getFullYear() &&
+              Number(booking.date.substring(5, 7)) === date.getMonth() + 1 &&
+              Number(booking.date.substring(8, 10)) === date.getDate() &&
+              booking.timeslotID === selected.id
+            ) {
+              const user = await DataStore.query(User, booking.userID);
+              if (user) {
+                if (user.userType === "Volunteer") {
+                  volUsers.push(user);
+                } else if (user.userType === "Rider") {
+                  ridUsers.push(user);
+                }
+              }
+            }
+          }
+        }
+      }
+      return { volUsers, ridUsers };
+    };
+    const pullData = async () => {
+      if (selected) {
+        const volBookingsArray = await selected.volunteerBookings.toArray(); // turns out the volunteer and rider booking arrays
+        // in our objects just return the same thing so there's not really a point to them
+        const bookings = await getUsers(volBookingsArray);
+        setVolBookings(bookings.volUsers);
+        setRidBookings(bookings.ridUsers);
+      }
+    };
     fetchBookable();
+    pullData();
   }, [popup]);
 
   return (
@@ -183,13 +219,10 @@ export default function Popup({
             <Wrapper>
               <LeftColumn>
                 <Monthly />
-                {selected && (
-                  <AptInfo
-                    timeslot={selected}
-                    date={date}
-                    // setTimeslot={setSelected}
-                  />
-                )}
+                <AptInfo
+                  riderBookings={riderBookings}
+                  volunteerBookings={volunteerBookings}
+                />
               </LeftColumn>
               <RightColumn>
                 <DateHeader>{formattedDate}</DateHeader>
